Sample framework for testing and such
This is the first pass at a framework for testing, using cmake
with files shamelessly stolen from Dr. Sutton. With the correct
programs installed it should also naturally lend to printing of
code in an acceptable manner. Needs some work, will continue
working in this "sample" directory to try to build a "style-guide"
idea as well as expanding on the framework to make it more
flexibile.