Modified test framework, fixed up sample
The test framework now has a diff function explicitly for problems.
That is, if you're working on a problem like Sweet Tarts and you
follow the pattern of the sample, you can just use the diff
function with apropriate input and output files, like a normal
test for a problem set.